🔄🎨 Change UI of podium and move section to main
