+++ src/app/(main)/constants/constants.tsx
@@ -1,0 +1,9 @@
+export const sportsIcons = [
+  { src: '/images/บอล.png', alt: 'Football' },
+  { src: '/images/บาส.png', alt: 'Basketball' },
+  { src: '/images/แบตมินตัน.png', alt: 'Badminton' },
+  { src: '/images/ปิงปอง.png', alt: 'Table Tennis' },
+  { src: '/images/กรีฑา.png', alt: 'Running' },
+  { src: '/images/พื้นบ้าน.png', alt: 'Traditional sports' },
+  { src: '/images/Esport.png', alt: 'Esport' }
+]
